refactor(room): extract useRoomSettings from useRoomScreenState (C13)
New hook useRoomSettings owns settingsSheetVisible, bgmEnabled,
and the four open/close/animationChange/bgmChange callbacks.
useRoomScreenState delegates to useRoomSettings.
